[Refactor] : noticeBoard couldn't to be deleted #625
생성 요청을 못하기 때문에, 삭제를 할 수 없습니다.

+++ src/main/java/peer/backend/service/board/team/BoardService.java
@@ -153,6 +153,9 @@ public class BoardService {
         if (!teamService.isLeader(board.getTeam().getId(), user)) {
             throw new ForbiddenException("팀을 삭제할 권한이 없습니다.");
         }
+        if (board.getType().equals(BoardType.NOTICE)) {
+            throw new ForbiddenException("공지사항 게시판은 삭제할 수 없습니다.");
+        }
         boardRepository.delete(board);
     }
